+++ CHANGELOG.md
@@ -3,7 +3,7 @@ aws-parallelcluster-cookbook CHANGELOG
 
 This file is used to list changes made in each version of the AWS ParallelCluster cookbook.
 
-3.X.X
+3.1.0
 ------
 
 **ENHANCEMENTS**